web-app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
redis==5.0.4
boto3==1.34.96
pydantic==2.7.1
